requirements.txt
mlx>=0.11
huggingface-hub
Pillow
matplotlib
gradio
tqdm
regex